backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0        
python-dotenv==1.0.1
google-genai
pydantic==2.6.4
